Alteração: mudança do tipo do Id(Long)

+++ src/main/java/org/example/dto/CategoryDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class CategoryDTO {
-    private int id;
+    private Long id;
     private String name;
-    private int lineId;
+    private Long lineId;
 }
