+++ src/main/java/org/example/dto/ProductDTO.java
@@ -4,7 +4,7 @@ import lombok.Data;
 
 @Data
 public class ProductDTO {
-    private int id;
+    private Long id;
     private String name;
-    private int categoryId;
+    private Long categoryId;
 }
