+++ src/main/java/org/example/repositories/CategoryRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface CategoryRepository extends JpaRepository<Category, Integer> {
-    List<Category> findByLineId(int lineId);
+public interface CategoryRepository extends JpaRepository<Category, Long> {
+    List<Category> findByLineId(Long lineId);
 }
